Modify /etc/hosts with trusted peers
When a WireHub network is up, trusted peers with an hostname are added
to /etc/hosts to make the system able to resolve WireHub peers hostname.

This approach may be unsecure, as it modifiy /etc/hosts. Race conditions
are possible. It is not advised to enable this feature outside a
container.

For now, this feature is by default disabled. To enable it, set the
environment variable `WH_EXPERIMENTAL_MODIFY_HOSTS` to `true`. The
Docker sandbox has the environment variable set.

NOTE: other approaches were considered, like NSS, which appears to not
be easily supported in Alpine Linux.