Add configuration reload capability
The method n.reload forces the node to update its peer's table with its
initial configuration which might have changed. New aliases and new
trusted peers are added, old aliases are removed and old trusted peers
are set as not trusted anymore.

The IPC 'reload' and CLI command 'reload' are added to force a daemon to
reload remotely.