Simpler configuration management
Previously, wh provided a way to manage configuration of networks. The
format of this configuration were similar to WireGuard's.

After using wh, I think a simpler-yet-complete description of a network
should be only a list of hostname and public key for each node. Like so,

    name test
    subnet 10.0.42.0/24
    workbit 8

    boot P17zMwXJFbBdJEn05RFIMADw9TX5_m2xgf31OgNKX3w bootstrap.wirehub.io

    trust root.test 10.0.42.1 KJ7YGrBeqLLm_JJ1huIS26OnqAVFy57z5UJqjyMagW4
    trust b.test              kKZzuIm11zkBSHL9ETRwEthIBbLTvz840F_k4mhI_Hs
    alias c.test              eIix5ldvqDzOIrG9ViKTe9TSBlF4g9nUwKi20C06hFM
    ...

The syntax format is described here:

    LINE := {
        PEER [<hostname>] [<private ip>] <public_key> [<endpoint>] |
        'name' <network name> |
        'namespace' <network namespace> |
        'subnet' <subnet ip>/<subnet cidr> |
        'workbit' <workbit>
    }

    PEER := { 'alias', 'boot', 'peer', 'router', 'trust' }

This patch implements this new configuration format. It also removes all
subcommands related to setting and getting network configurations. Now,
a network configuration is a single file passed as an argument to
wh. The basic wh commands is now

    wh up ./config private-key ./sk